move components to components/
