+++ client/src/components/SetContent-Album.jsx
