+++ client/src/components/UserBadge.jsx
